requirements.txt
fastapi
uvicorn[standard]
pymongo
httpx
Jinja2
python-dotenv
pydantic-settings
